requirements/base.txt
Django==2.0.6
djangorestframework==3.8.2
jsonfield==2.0.2
pytz==2018.5
